Moves JobItem to a separate component
