+++ src/components/JobsList/JobsList.jsx
@@ -7,25 +7,9 @@ import {
   IonLabel,
   IonList
 } from '@ionic/react';
-import { magnetOutline } from 'ionicons/icons';
 import { useSelector } from 'react-redux';
+import JobItem from '../JobItems/JobItem';
 
-export const JobItem = ( job, key ) => {
-  return (
-    <IonItem key={key} onClick={handleClick}>
-      <IonIcon icon={magnetOutline} slot="start" />
-      <IonLabel slot=''>{job.company}</IonLabel>
-      <IonLabel slot=''>Some job info here</IonLabel>
-      <IonLabel slot=''>Salary</IonLabel>
-      <IonLabel slot=''>Status</IonLabel>
-      <IonChip slot="end" >99</IonChip>
-    </IonItem>
-  );
-}
-
-function handleClick({ target }) {
-  console.log(`The index of this object in the jobs array is ${JSON.stringify(target.value)}`);
-}
 
 const JobsList = ({jobs}) => {
   // let jobState = useSelector((state) => state.jobs.jobs);
